notices controller works fine

+++ asta-bank-security/src/main/java/com/spring/astabanksecurity/repository/NoticeRepository.java
@@ -8,7 +8,7 @@ import org.springframework.stereotype.Repository;
 import java.util.List;
 
 @Repository
-public interface NoticeRepository extends JpaRepository<Notice, Long> {
+public interface NoticeRepository extends JpaRepository<Notice, Integer> {
 
     @Query(value = "from Notice n where CURDATE() BETWEEN noticeBeginDate and noticeEndDate")
     List<Notice> findAllActiveNotices();
